Created separate file for window and doorway functions
The functions were also renamed to not contain 2DES as these function could be used for other techniques.
