+++ src/CMakeLists.txt
@@ -17,7 +17,7 @@ add_executable(NISE
     analyse.h readinput.c readinput.h propagate.c propagate.h
     read_trajectory.c read_trajectory.h correlate.h correlate.c
     project.c project.h types_MPI.h types_MPI.c mcfret.c mcfret.h
-    calc_Redfield.c calc_Redfield.h 
+    calc_Redfield.c calc_Redfield.h CG_subs.c CG_subs.h
     sfg.c sfg.h calc_CG_2DES.c calc_CG_2DES.h eq_den.c eq_den.h
     calc_FD_CG_2DES.c calc_FD_CG_2DES.h   
     $<TARGET_OBJECTS:random_lib>
